// apps/front/src/redux/Slice/useSliceForAdmin/user.types.ts
// redux/types/user.types.ts

export type UserRole = "superadmin" | "admin" | "user";

export interface IUser {
  _id: string;
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  role: UserRole;
  isVerified: boolean;
  isActive: boolean;
  // add other fields if needed
}

export interface UserState {
  user: IUser | null;
  users: IUser[];
  loading: boolean;
  error: string | null;
}